Fix stopwatch format
