+++ .expo/web/cache/development/babel-loader/3ddc6c966e7b433144d68de019931e47ca70c3bd39b1beb07b3d0819215cf322.json
@@ -1,0 +1,1 @@
+{"ast":null,"code":"import StyleSheet from \"react-native-web/dist/exports/StyleSheet\";\nimport View from \"react-native-web/dist/exports/View\";\nimport StopWatch from \"./components/StopWatch\";\nimport { jsx as _jsx } from \"react/jsx-runtime\";\nexport default function App() {\n  return _jsx(View, {\n    style: styles.container,\n    children: _jsx(StopWatch, {})\n  });\n}\nvar styles = StyleSheet.create({\n  container: {\n    flex: 1,\n    backgroundColor: '#fff',\n    alignItems: 'center',\n    justifyContent: 'center'\n  }\n});","map":{"version":3,"names":["StopWatch","jsx","_jsx","App","View","style","styles","container","children","StyleSheet","create","flex","backgroundColor","alignItems","justifyContent"],"sources":["C:/Users/pedro/Documents/FitQuest/semester-project-fitquest-gameified-fitness-application/Front End/App.tsx"],"sourcesContent":["import { StatusBar } from 'expo-status-bar';\r\nimport { StyleSheet, Text, View } from 'react-native';\r\nimport StopWatch from './components/StopWatch';\r\n\r\nexport default function App() {\r\n  return (\r\n    <View style={styles.container}>\r\n      <StopWatch/>\r\n    </View>\r\n  );\r\n}\r\n\r\nconst styles = StyleSheet.create({\r\n  container: {\r\n    flex: 1,\r\n    backgroundColor: '#fff',\r\n    alignItems: 'center',\r\n    justifyContent: 'center',\r\n  },\r\n});\r\n"],"mappings":";;AAEA,OAAOA,SAAS;AAA+B,SAAAC,GAAA,IAAAC,IAAA;AAE/C,eAAe,SAASC,GAAGA,CAAA,EAAG;EAC5B,OACED,IAAA,CAACE,IAAI;IAACC,KAAK,EAAEC,MAAM,CAACC,SAAU;IAAAC,QAAA,EAC5BN,IAAA,CAACF,SAAS,IAAC;EAAC,CACR,CAAC;AAEX;AAEA,IAAMM,MAAM,GAAGG,UAAU,CAACC,MAAM,CAAC;EAC/BH,SAAS,EAAE;IACTI,IAAI,EAAE,CAAC;IACPC,eAAe,EAAE,MAAM;IACvBC,UAAU,EAAE,QAAQ;IACpBC,cAAc,EAAE;EAClB;AACF,CAAC,CAAC"},"metadata":{},"sourceType":"module","externalDependencies":[]}
